% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{bars}
\alias{bars}
\title{Data frame for an example bar chart}
\usage{
bars(n)
}
\arguments{
\item{n}{The number of categories in column \code{col2}, which is intended to be
mapped to the fill aesthetic of the plot. Must be from one to six.}
}
\value{
Returns a data frame with the specified number of categories in \code{col2}.
}
\description{
Creates a data frame that can be used for an example bar chart.
}
\examples{
bars(3)
}
