% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_colours}
\alias{display_colours}
\title{Display a colour palette}
\usage{
display_colours(x)
}
\arguments{
\item{x}{A character vector of colours.}
}
\description{
Displays a vector of colours.
}
\examples{
display_colours(c("#039BE5", "#E91E63", "#FFB300"))
display_colours(mojchart_palette(4, palette = "muted1"))
display_colours(moj_colours())
}
