% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_palette}
\alias{display_palette}
\title{Display a colour scheme}
\usage{
display_palette(palette)
}
\arguments{
\item{palette}{The name of an mojchart colour scheme. Run \code{palette_names()} for
the available options.}
}
\description{
Preview all the palettes contained within an mojchart colour scheme.
}
\examples{
display_palette("vibrant1")
}
