% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_scheme}
\alias{display_scheme}
\title{Display a colour scheme}
\usage{
display_scheme(scheme)
}
\arguments{
\item{scheme}{The name of an mojchart colour scheme. Run \code{scheme_names()} for
the available options.}
}
\description{
Preview all the palettes contained within an mojchart colour scheme.
}
\examples{
display_scheme("vibrant1")
}
