% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{familystarts}
\alias{familystarts}
\title{Time series of case starts in the family court (published statistics)}
\format{
A data frame with 39 rows and 11 Variables: \describe{
\item{year_qtr}{Year and quarter}
\item{case_type}{Case type}
\item{count}{Number of cases}
}
}
\source{
Family Court Statistics Quarterly, April to June 2020\cr
\url{https://www.gov.uk/government/statistics/family-court-statistics-quarterly-april-to-june-2020}
}
\usage{
familystarts
}
\description{
A time series of the number of cases started in the family court in England
and Wales, from 2011 to 2020, broken down by case type. Intended for example
line charts with one to six lines.
}
\keyword{datasets}
