% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{familystarts_reduced}
\alias{familystarts_reduced}
\title{Reduce familystarts}
\usage{
familystarts_reduced(n = 6)
}
\arguments{
\item{n}{The number of case types required, from one to six.}
}
\value{
Returns the filtered data frame.
}
\description{
Reduces the familystarts dataset to a given number of case types, from one to
six. Intended for creating example line charts.
}
\examples{
familystarts_reduced(4)
}
