% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{govanal_colours}
\alias{govanal_colours}
\title{Government Analysis Function colours}
\usage{
govanal_colours(...)
}
\arguments{
\item{...}{Colour names as character strings.}
}
\value{
Returns a named character vector.
}
\description{
Provides names and hex codes for the Government Analysis Function colours.
Calling the function with no arguments returns every colour. If the names of
specific colours are passed as arguments, the function returns those colours
alone. See \url{https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/}
}
\examples{
# All colours
govanal_colours()

# Selected colours
govanal_colours("govanal_darkblue", "govanal_orange")
}
\seealso{
Other available colours: 
\code{\link{moj_colours}()},
\code{\link{palette_colours}()}
}
\concept{available colours}
