% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{moj_colours}
\alias{moj_colours}
\title{MoJ colours}
\usage{
moj_colours(...)
}
\arguments{
\item{...}{Colour names as character strings.}
}
\value{
Returns a named character vector.
}
\description{
Provides names and hex codes for MoJ corporate colours. Calling the function
with no arguments returns every colour. If the names of specific colours are
passed as arguments, the function returns those colours alone.
}
\examples{
# All colours
moj_colours()

# Selected colours
moj_colours("mojblue", "mojgreen")
}
\seealso{
MoJ branding intranet page:
\url{https://intranet.justice.gov.uk/guidance/communications/branding-templates/}

Other available colours: 
\code{\link{govanal_colours}()},
\code{\link{palette_colours}()}
}
\concept{available colours}
