% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{moj_palette}
\alias{moj_palette}
\title{mojchart colour palette}
\usage{
moj_palette(n, palette)
}
\arguments{
\item{n}{The required number of colours in the palette, from one to six.}

\item{palette}{The name of an mojchart palette as a character string. Run
\code{moj_palette_names()} for the available options.}
}
\description{
Returns an mojchart colour palette as a named character vector.
}
\examples{
moj_palette(5, palette = "vibrant1")
}
