% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{multiswatch}
\alias{multiswatch}
\title{Display a colour scheme}
\usage{
multiswatch(...)
}
\arguments{
\item{...}{Arguments passed to \code{display_scheme()}.}
}
\description{
RENAMED: please use \code{display_scheme()} instead.
}
\examples{
multiswatch("vibrant1")
}
