% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{palette_colours}
\alias{palette_colours}
\title{Palette colours}
\usage{
palette_colours(...)
}
\arguments{
\item{...}{Colour names as character strings.}
}
\value{
Returns a named character vector.
}
\description{
Provides names and hex codes for colours used in the mojchart palettes. If
colour names are passed as arguments, returns those colours. If no arguments
are provided, returns every colour.
}
\examples{
# All colours
palette_colours()

# Selected colours
palette_colours("mojblue", "yellow1")
}
\seealso{
Other available colours: 
\code{\link{govanal_colours}()},
\code{\link{moj_colours}()}
}
\concept{available colours}
