% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{swatch}
\alias{swatch}
\title{Display a colour palette}
\usage{
swatch(...)
}
\arguments{
\item{...}{Arguments passed to \code{display_palette()}.}
}
\description{
RENAMED: please use \code{display_palette()} instead.
}
\examples{
swatch(moj_colours())
}
