% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_palette}
\alias{display_palette}
\title{Display a colour palette}
\usage{
display_palette(x)
}
\arguments{
\item{x}{A character vector of colours.}
}
\description{
Displays a vector of colours.
}
\examples{
display_palette(c("#039BE5", "#E91E63", "#FFB300"))
display_palette(mojchart_palette(4, scheme = "muted1"))
display_palette(moj_colours())
}
