% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{moj_palette}
\alias{moj_palette}
\title{Colour palette}
\usage{
moj_palette(n, palette)
}
\arguments{
\item{n}{The number of colours required, from one to six.}

\item{palette}{The name of a colour palette as a character string. One of
govanal_bars, govanal_lines, muted1, muted2, muted3, vibrant1, vibrant2}
}
\description{
Returns one of the package colour palettes as a named character vector.
}
\examples{
moj_palette(5, palette = "vibrant1")
}
