% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{mojchart_palette}
\alias{mojchart_palette}
\title{Colour palette}
\usage{
mojchart_palette(...)
}
\arguments{
\item{...}{Arguments passed to \code{moj_palette()}.}
}
\description{
RENAMED: please use \code{moj_palette()} instead.
}
\examples{
mojchart_palette(5, palette = "vibrant1")
}
