% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{mojchart_palette}
\alias{mojchart_palette}
\title{Colour palette}
\usage{
mojchart_palette(n, scheme)
}
\arguments{
\item{n}{The number of colours required, from one to six.}

\item{scheme}{The name of a colour scheme as a character string. Run
\code{scheme_names()} for the available options.}
}
\description{
Returns one of the package colour palettes as a named character vector.
}
\examples{
mojchart_palette(5, scheme = "vibrant1")
}
