% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatch.R
\name{multiswatch}
\alias{multiswatch}
\title{Swatch plot for a full colour scheme}
\usage{
multiswatch(scheme)
}
\arguments{
\item{scheme}{The name of an mojchart colour scheme. Run \code{scheme_names()} for
the available options.}
}
\description{
Preview all the palettes contained within an mojchart colour scheme.
}
\examples{
multiswatch("vibrant1")
}
