% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatch.R
\name{multiswatch}
\alias{multiswatch}
\title{Swatch plot for a full colour scheme}
\usage{
multiswatch(palette_name)
}
\arguments{
\item{palette_name}{The name of an mojchart palette. Run
\code{moj_palette_names()} for the available options.}
}
\description{
Preview all the palettes contained within an mojchart colour scheme.
}
\examples{
multiswatch("vibrant1")
}
