% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatch.R
\name{swatch}
\alias{swatch}
\title{Swatch plot}
\usage{
swatch(colours)
}
\arguments{
\item{colours}{A character vector of colours.}
}
\description{
Preview a vector of colours.
}
\examples{
swatch(c("#039BE5", "#E91E63", "#FFB300"))
swatch(mojchart_palette(4, scheme = "muted1"))
swatch(moj_colours())
}
