% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gss.R
\name{theme_gss}
\alias{theme_gss}
\title{GSS theme}
\usage{
theme_gss(base_size = 20, base_line_size = 0.5, flipped = FALSE)
}
\arguments{
\item{base_size}{The base font size for chart text. The default is set to
size 20, which works for publication charts on gov.uk. (The ggplot2 default
size is 11.) For further details, see
\url{https://ggplot2-book.org/polishing.html}.}

\item{base_line_size}{The base size for lines controlled by the theme
function (i.e. non-data lines).}

\item{flipped}{A logical value indicating whether the chart has flipped axes.}
}
\value{
Returns a ggplot2 theme function.
}
\description{
A theme based on Government Statistical Service (GSS) guidance.\cr
\url{https://gss.civilservice.gov.uk/policy-store/introduction-to-data-visualisation/}
}
\examples{
library(ggplot2)
ggplot(familystarts, aes(x = year_qtr, y = count, colour = case_type)) +
  geom_line(size = 1.5) +
  labs(title = "Number of cases started in family courts",
       subtitle = "England and Wales, Q1 2011 - Q2 2020",
       caption = "Source: Family court statistics quarterly, April to June 2020 (table 1)") +
  theme_gss(base_size = 18)
}
